Extract dashboard management tests to a separate spec
